Fix NEAT output problem; Fix SpelunkBots bug
This commit fixes a weird SpelunkBots bug where the bots would carry out
commands from previously ran bots for a couple of frames. This led to undesired
behaviors that varied from small hops in the beginning of the run to nasty
exploits like acquiring an enormous speed boost.

To fix this, I simply changed the order that a few scripts executed inside the
GameMaker project. In the object "oPlayer"'s step event, I moved the script
"ResetPlayerCommands" to be executed earlier. Before, it was executed right
after the "UpdateGlobalAIVariables" script. The order now is:

* Code
* Action (comment)
* ResetPlayerCommands
* Code
* Code
* ...