Add SpelunkBots configuration file
When we wanted to change SpelunkBots level parameters, we were forced to do so
inside the Game Maker project. That was annoying, because that meant we had to
save the project and export the executable again. That also meant there was no
way for git to add these modifications to our changelogs (binary files).

We are now able to set these parameters through an .INI file, which Game Maker
supports as a way of reading parameters from external files. The SpelunkBots INI
file is located at "Source/spelunky_1_1", same directory as the executable.

This commit also fixes a terrible bug on SpelunkBots related to the level names.
SpelunkBots was using a vector of pointers to chars (vector<char*>) to hold
level names. When we used the INI files to read level names and, later during
the game's execution, requested the level name through "CheckNextLevel",
SpelunkBots was returning garbage. Also, the framework was using two separate
vectors to hold level names and seeds, which was completely unecessary.